+++ test/engine/engine.test.ts
@@ -308,6 +308,7 @@ describe('Engine', function () {
 
     builtinTokens: {},
     currencyInfo: {} as any,
+    defaultNetworkInfo: {},
     networkInfo: {}
   }